CORE: - Fix NumericByteArraySummaryType to return a byte instead of a long.   Copy-pasta bug. - Change Tsdb1xPartialTimeSeries to be a base class with default   implementations. - Rework TestTsdb1xNumericPartialTimeSeries with the new base class and   have it re-grow the array instead of creating new ReAllocatedArray's   every time.

+++ core/src/main/java/net/opentsdb/data/types/numeric/NumericByteArraySummaryType.java
@@ -70,6 +70,6 @@ public interface NumericByteArraySummaryType extends TimeSeriesDataType {
    * {@link offset()} must equal {@link end()}.
    * @return The array of data, may be null or empty.
    */
-  public long[] data();
+  public byte[] data();
   
 }
